Moved cookies stuff into separate file
